# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 途个惊喜 API
  version: 1.0.0
  description: |
    用户端「原生微信小程序」与管理端「Vue3 Web」共用的业务 API。

    ## 约定
    - 用户端前缀 `/api/v1`，标注「需登录」的接口带用户 JWT。
    - 管理端前缀 `/api/v1/admin`，需管理员 JWT + 角色校验。
    - 统一响应体 `{ code, message, data }`，`code=0` 为成功。
    - **金额在接口层一律用「元」**（最多两位小数，`number` 类型）；
      数据库内部用「分」（`INT UNSIGNED`），换算只发生在适配层。
    - **时间一律用字符串**，格式 `yyyy-MM-dd HH:mm:ss`（北京时间，不用 UTC）。
    - 列表统一返回 `data: { list, total, page, pageSize }`。

    ## 与文档的关系
    本文件是接口契约的唯一真相（single source of truth）。
    `docs/接口设计.md` 是它的中文说明，两者必须同步修改。
    后端用 springdoc 从代码生成后与本文件比对，避免漂移。
servers:
  - url: http://localhost:8080
    description: 本地开发
...
security:
  - UserToken: []
  - AdminToken: []
paths:
...
  /api/v1/admin/users:
    get:
      tags: [管理端-订单支付用户]
      summary: 用户列表
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
            description: 昵称 / 手机号 / 微信昵称
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled]
        - name: channel
          in: query
          schema:
            type: string
            enum: [phone, wechat]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 用户分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAdminUserResult'
...
  /api/v1/admin/users/{id}:
    get:
      tags: [管理端-订单支付用户]
      summary: 用户完整档案
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200':
          description: 资料、统计、徽章、订单、行程与客服信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AdminUserDetail' }
  /api/v1/admin/users/{id}/status:
    patch:
      tags: [管理端-订单支付用户]
      summary: 禁用或恢复用户
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, disabled] }
                reason: { type: string, maxLength: 255 }
      responses:
        '200':
          description: 已切换
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResultVoid' }
  /api/v1/admin/users/{id}/note:
    patch:
      tags: [管理端-订单支付用户]
      summary: 修改客服备注
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: 已保存
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResultVoid' }
...
components:
  securitySchemes:
    UserToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户端 JWT，前缀 /api/v1
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 管理端 JWT，前缀 /api/v1/admin
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
    ResultVoid:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data:
          type: object
          nullable: true
...
    Category:
      type: string
      description: 周边游 / 省内游 / 跨省游 / 主题专线
      enum: [nearby, province, cross, theme]
    Mood:
      type: string
      description: 开心 / emo / 无聊 / 迷茫
      enum: [happy, emo, bored, curious]

    # ── 用户端 ──
...
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 古村
        mark:
          type: string
          description: 印章内的单字
          example: 村
        description:
          type: string
        unlocked:
          type: boolean
        unlockedAt:
          type: string
          nullable: true
...
    UserStats:
      type: object
      description: 口径见数据库设计 §5
      properties:
        tripCount:
          type: integer
          description: 有效行程数
        savedTotal:
          type: number
          description: 累计省钱 = Σ(票面 − 购入价)，仅有效行程
        levelNo:
          type: integer
          description: 当前等级数字，max(1, tripCount)
        title:
          type: string
        welfareKm:
          type: number
          description: 公益里程，按有效行程累计
        badgeUnlocked:
          type: integer
        badgeTotal:
          type: integer
        personalityType:
          type: string
          nullable: true
        destinations:
          type: array
          items:
            type: string
        destinationCount:
          type: integer
          description: 有效行程目的地去重数量
        spendTotal:
          type: number
          description: 累计消费，仅支付成功且未退款
        recentMood:
          $ref: '#/components/schemas/Mood'
          nullable: true
        favoriteCategory:
          $ref: '#/components/schemas/Category'
          nullable: true
        monthTrips:
          type: integer
        diaryCount:
          type: integer

    # ── 管理端 ──
...
    AdminOrder:
      type: object
      properties:
        orderNo:
          type: string
        userName:
          type: string
        phone:
          type: string
          description: 脱敏
        boxName:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
          nullable: true
        channel:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending_pay, paid, opened, cancelled, refunded]
        routeName:
          type: string
          nullable: true
        createdAt:
          type: string
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
        phone:
          type: string
        wechatNickname:
          type: string
          nullable: true
        channel:
          type: string
          enum: [phone, wechat]
        tripCount:
          type: integer
        title:
          type: string
        spendTotal:
          type: number
        savedTotal:
          type: number
        status:
          type: string
          enum: [active, disabled]
        personalityType:
          type: string
          nullable: true
        lastLoginAt:
          type: string
          nullable: true
        createdAt:
          type: string
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/AdminUser'
        - type: object
          properties:
            recentMood:
              $ref: '#/components/schemas/Mood'
              nullable: true
            city:
              type: string
              nullable: true
            gender:
              type: integer
            stats:
              $ref: '#/components/schemas/UserStats'
            badges:
              type: array
              items:
                $ref: '#/components/schemas/Badge'
            recentOrders:
              type: array
              items:
                $ref: '#/components/schemas/AdminOrder'
            csNote:
              type: string
              nullable: true
            disableReason:
              type: string
              nullable: true
...
    PagedAdminUserResult:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminUser'
...
    PageMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer